Rename publish executor files
Align executor folder and file names with the publish executor key while keeping the target name as nx-release-publish, and update plan/test references to avoid leftover mismatch.
