+++ packages/nx-terraform-plugin/src/__tests__/project.test.ts
@@ -64,7 +64,7 @@ const getExpectedDocsTarget = () => ({
 
 const getExpectedPublishTarget = () => ({
   cache: false,
-  executor: "@pagopa/nx-terraform-plugin:nx-release-publish",
+  executor: "@pagopa/nx-terraform-plugin:publish",
   options: {
     projectRoot: "{projectRoot}",
     workspaceRoot: "{workspaceRoot}",
