+++ packages/nx-terraform-plugin/src/executors/publish/schema.json
@@ -1,7 +1,7 @@
 {
   "$schema": "https://json-schema.org/schema",
   "version": 2,
-  "title": "NxReleasePublish executor",
+  "title": "Publish executor",
   "description": "",
   "type": "object",
   "properties": {
